+++ lib/hashie/extensions/pretty_inspect.rb
@@ -8,7 +8,7 @@ module Hashie
 
       def hashie_inspect
         ret = "#<#{self.class}"
-        stringify_keys.keys.sort.each do |key|
+        keys.sort_by { |key| key.to_s }.each do |key|
           ret << " #{key}=#{self[key].inspect}"
         end
         ret << '>'